+++ packages/nodepay-ezidebit/src/types/payment-schedule.ts
@@ -1,5 +1,24 @@
-import { EzidebitDayOfWeek } from './payment-frequency'
+import { EzidebitDayOfWeek, EzidebitPaymentFrequency } from './payment-frequency'
 import { IPaymentSchedule } from '@atomixdesign/nodepay-core/types'
+
+/** @internal */
+export interface IEzidebitInternalPaymentSchedule {
+  ScheduleStartDate: string
+  SchedulePeriodType: EzidebitPaymentFrequency
+  DayOfMonth: number
+  PaymentAmountInCents: number
+  LimitToNumberOfPayments: number
+  LimitToTotalAmountInCents: number
+  KeepManualPayments: string
+  EziDebitCustomerID: string
+  YourSystemReference: string
+  DayOfWeek: EzidebitDayOfWeek
+  FirstWeekOfMonth: string
+  SecondWeekOfMonth: string
+  ThirdWeekOfMonth: string
+  FourthWeekOfMonth: string
+  Username: string
+}
 
 export interface IEzidebitPaymentSchedule extends IPaymentSchedule {
   ezidebitCustomerId: string
