+++ packages/nodepay-ezidebit/src/types/payment.ts
@@ -1,5 +1,3 @@
-import { EzidebitPaymentFrequency, EzidebitDayOfWeek } from './payment-frequency'
-
 export interface IEzidebitBasePayment {
   PaymentAmountInCents: number
   PaymentReference: string
@@ -7,24 +5,6 @@ export interface IEzidebitBasePayment {
 
 export interface IEzidebitCustomerName {
   CustomerName: string
-}
-
-export interface IEzidebitInternalPaymentSchedule {
-  ScheduleStartDate: string
-  SchedulePeriodType: EzidebitPaymentFrequency
-  DayOfMonth: number
-  PaymentAmountInCents: number
-  LimitToNumberOfPayments: number
-  LimitToTotalAmountInCents: number
-  KeepManualPayments: string
-  EziDebitCustomerID: string
-  YourSystemReference: string
-  DayOfWeek: EzidebitDayOfWeek
-  FirstWeekOfMonth: string
-  SecondWeekOfMonth: string
-  ThirdWeekOfMonth: string
-  FourthWeekOfMonth: string
-  Username: string
 }
